Fix Google layer showing around OS layer
The Google layers do not heed the functions to set base layer until
they have been loaded. Because the OS maps have limited extents,
the Google layer could show up around the edges. To prevent this, a
callback is added to the loading of Google layers. It ensures they
are hidden if not the baselayer.